Subscription renewal, cleanup 2.0 Added: - util.py (instead of multiple smaller files) - Dockerfile, sleep infinity option for testing - subscription auto renewal (should not stop after a day) - if name = main check
Changed:
- Listener in function
- Cleaner printing
- More code cleanup
- Quicker volume syncing

Removed:
- active.py, volume.py, discover.py
- Unused modules import
